requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
scikit-learn>=1.3.0

# LLM Provider Support

# Ollama support
ollama>=0.1.0

# Gemini support (Google GenAI SDK)
google-genai>=0.2.0

# OpenAI support
openai>=1.0.0
tiktoken>=0.5.0

# Additional dependencies
pathlib>=1.0.1  # For path handling (usually included in Python 3.4+)
typing-extensions>=4.0.0  # For extended type hints
